+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+- Added metrics diagnostics
+- Added `WP_Object_Cache::decr()` alias
+- Moved `diagnostics.php` file
+
 ## 2.0.20
 
 - Fix wp.org release